+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add a `--substitute-on-destination` argument.
 - Add the `meta.nodeSpecialArgs` meta option.
+- Add `wire build`, a new command to build nodes offline.
+  It is distinct from `wire apply build`, as it will not ping
+  or push the result, making it useful for CI.
 
 ## [v1.0.0] - 2025-12-17